+++ src/Schema/SchemaInterface.php
@@ -11,6 +11,7 @@ use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Schema\Builder\ColumnInterface;
+use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 
 /**
  * Represents the schema for a database table.
@@ -227,6 +228,11 @@ interface SchemaInterface extends ConstraintSchemaInterface
      * @psalm-param string[]|int[]|int|string|null $length
      */
     public function createColumn(string $type, array|int|string $length = null): ColumnInterface;
+
+    /**
+     * Returns the column factory for creating column instances.
+     */
+    public function getColumnFactory(): ColumnFactoryInterface;
 
     /**
      * @return string|null The default schema name.