+++ tests/Db/Schema/ColumnSchemaTest.php
@@ -130,6 +130,16 @@ final class ColumnSchemaTest extends TestCase
         $this->assertSame('', $column->getExtra());
     }
 
+    /** @dataProvider \Yiisoft\Db\Tests\Provider\ColumnSchemaProvider::load */
+    public function testLoad(string $parameter, mixed $value, string $method, mixed $expected): void
+    {
+        $column = new ColumnSchema();
+
+        $column->load([$parameter => $value]);
+
+        $this->assertSame($expected, $column->$method());
+    }
+
     public function testName(): void
     {
         $column = new ColumnSchema();